Reinstall migrations, load filemetadatastatus enum in migrations
