+++ .env.example
@@ -1,6 +1,3 @@
-# DOCKER
-DOCKER_DEPLOY=1
-
 # MOD
 DEBUG=1
