+++ src/files/models.py
@@ -1,19 +1,9 @@
 from django.db import models
 from django.conf import settings
 
-from enum import Enum
+from .enums import FileMetadataStatusEnum
+
 import uuid
-
-
-class FileMetadataStatusEnum(Enum):
-    """
-    Перечисления для статуса FileMetadata
-    формат: title = description
-    """
-    untracked = 'File not tracked on storage backend'
-    loaded = 'File load and exsist on storage backend'
-    deleted = 'File exsist, but not tracked on storage backend'
-    on_error = 'Some error with file (probably missing on storge backend)'
 
 
 class BaseModel(models.Model):
